+++ frameworks/gson-bench/src/main/java/com/github/skjolber/nve/gson/ReducedGsonParser.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public class ReducedGsonParser {
 
-    public static int countDefCveItem(List<byte[]> contents) throws IOException {
+    public static long countDefCveItem(List<byte[]> contents) throws IOException {
 
-		int count = 0;
+    	long count = 0;
 		
 		for(byte[] content : contents) {
 			try (GsonParser<org.nvd.json.gson.reduced.DefCveItem> reader = new GsonParser<>(content, org.nvd.json.gson.reduced.DefCveItem.class)) {
